requirements.txt
streamlit
torch
torchvision
numpy
pillow
imageio